requirements.txt
discord.py
psycopg2~=2.9.9